backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
bcrypt==4.1.3
httpx==0.28.1
